+++ editor-api/src/main/groovy/ish/oncourse/willow/editor/model/MenuItem.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public class MenuItem  {
   
-    private Double id = null;
+    private Integer id = null;
     private String title = null;
     private String url = null;
     private Boolean expanded = null;
@@ -18,15 +18,15 @@ public class MenuItem  {
      * Unique identifier of menu item
      * @return id
      */
-    public Double getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Double id) {
+    public void setId(Integer id) {
        this.id = id;
     }
 
-    public MenuItem id(Double id) {
+    public MenuItem id(Integer id) {
       this.id = id;
       return this;
     }
